CIRC-844: Rename base class for aged to lost fee refunds.
