+++ src/test/java/api/loans/scenarios/OverrideRenewAgedToLostItemTest.java
@@ -1,0 +1,22 @@
+package api.loans.scenarios;
+
+import org.joda.time.DateTime;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import api.support.fixtures.AgeToLostFixture;
+import api.support.fixtures.OverrideRenewalFixture;
+
+public class OverrideRenewAgedToLostItemTest extends RefundAgedToLostFeesTestBase {
+  @Autowired
+  private OverrideRenewalFixture overrideRenewalFixture;
+
+  @Override
+  protected void performActionThatRequiresRefund(AgeToLostFixture.AgeToLostResult result,
+     DateTime actionDate) {
+
+    mockClockManagerToReturnFixedDateTime(actionDate);
+
+    overrideRenewalFixture.overrideRenewalByBarcode(result.getLoan(),
+      servicePointsFixture.cd1().getId());
+  }
+}
